# apps/web-site/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "企业级 AI Agent Harness API",
    "description": "用于企业级 AI Agent Harness 平台的任务、事件、沙箱、审计和设置 API。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/settings/models": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "查询模型设置",
        "description": "返回模型网关、供应商、限流和健康状态。",
        "operationId": "get_model_settings_api_settings_models_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelSettingsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "更新模型设置",
        "description": "仅 admin 可更新模型网关设置。",
        "operationId": "update_model_settings_api_settings_models_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelSettingsResponse"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelSettingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/settings/policies": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "查询策略设置",
        "description": "返回工具风险、审批规则、沙箱规则和审计要求。",
        "operationId": "get_policy_settings_api_settings_policies_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicySettingsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "更新策略设置",
        "description": "仅 admin 可更新策略设置。",
        "operationId": "update_policy_settings_api_settings_policies_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicySettingsResponse"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicySettingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/tasks/{task_id}/subagents": {
      "get": {
        "tags": [
          "subagents"
        ],
        "summary": "查询任务子 Agent",
        "description": "返回指定任务派生出的子 Agent 列表。",
        "operationId": "list_task_subagents_api_tasks__task_id__subagents_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubagentPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "subagents"
        ],
        "summary": "创建任务子 Agent",
        "description": "为指定任务派生一个子 Agent，并写入 SUBAGENT_SPAWNED 事件。",
        "operationId": "create_task_subagent_api_tasks__task_id__subagents_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubagentCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubagentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/subagents/{subagent_id}": {
      "get": {
        "tags": [
          "subagents"
        ],
        "summary": "查询子 Agent 详情",
        "description": "返回单个子 Agent 的状态与上下文。",
        "operationId": "get_subagent_api_subagents__subagent_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "subagent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subagent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubagentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/subagents/{subagent_id}/cancel": {
      "post": {
        "tags": [
          "subagents"
        ],
        "summary": "取消子 Agent",
        "description": "取消仍在运行或等待中的子 Agent。",
        "operationId": "cancel_subagent_api_subagents__subagent_id__cancel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "subagent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subagent Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubagentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelSettingsResponse": {
        "properties": {
          "default_provider": {
            "type": "string",
            "title": "Default Provider",
            "description": "默认供应商"
          },
          "default_model": {
            "type": "string",
            "title": "Default Model",
            "description": "默认模型"
          },
          "providers": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Providers",
            "description": "供应商列表"
          },
          "rate_limits": {
            "additionalProperties": true,
            "type": "object",
            "title": "Rate Limits",
            "description": "限流信息"
          },
          "health": {
            "additionalProperties": true,
            "type": "object",
            "title": "Health",
            "description": "健康状态"
          }
        },
        "type": "object",
        "required": [
          "default_provider",
          "default_model",
          "providers",
          "rate_limits",
          "health"
        ],
        "title": "ModelSettingsResponse"
      },
...
      "PolicySettingsResponse": {
        "properties": {
          "risk_levels": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Risk Levels",
            "description": "风险等级"
          },
          "approvals": {
            "additionalProperties": true,
            "type": "object",
            "title": "Approvals",
            "description": "审批规则"
          },
          "sandbox": {
            "additionalProperties": true,
            "type": "object",
            "title": "Sandbox",
            "description": "沙箱规则"
          },
          "audit": {
            "additionalProperties": true,
            "type": "object",
            "title": "Audit",
            "description": "审计规则"
          }
        },
        "type": "object",
        "required": [
          "risk_levels",
          "approvals",
          "sandbox",
          "audit"
        ],
        "title": "PolicySettingsResponse"
      },
...
      "SubagentCreateRequest": {
        "properties": {
          "assignment": {
            "additionalProperties": true,
            "type": "object",
            "title": "Assignment",
            "description": "子 Agent 任务上下文"
          },
          "parent_agent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Agent Id",
            "description": "父 Agent ID"
          },
          "timeout_seconds": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Timeout Seconds",
            "description": "超时秒数",
            "default": 900
          },
          "enqueue": {
            "type": "boolean",
            "title": "Enqueue",
            "description": "是否进入 Dramatiq 队列",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "assignment"
        ],
        "title": "SubagentCreateRequest"
      },
      "SubagentPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SubagentResponse"
            },
            "type": "array",
            "title": "Items",
            "description": "子 Agent 列表"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor",
            "description": "下一页游标"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "SubagentPage"
      },
      "SubagentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "子 Agent ID"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id",
            "description": "任务 ID"
          },
          "parent_agent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Agent Id",
            "description": "父 Agent ID"
          },
          "agent_type": {
            "type": "string",
            "title": "Agent Type",
            "description": "Agent 类型"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "子 Agent 状态"
          },
          "context_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Context Json",
            "description": "上下文"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At",
            "description": "开始时间"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At",
            "description": "完成时间"
          },
          "timeout_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout At",
            "description": "超时时间"
          }
        },
        "type": "object",
        "required": [
          "id",
          "task_id",
          "agent_type",
          "status",
          "context_json"
        ],
        "title": "SubagentResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}